+++ src/lib/st-form/st-form.module.ts
@@ -18,6 +18,7 @@ import { StInputModule } from '../st-input/st-input.module';
 import { StFormDirectiveModule } from '../directives/form/form-directives.module';
 import { StFormFieldModule } from './st-form-field/st-form-field.module';
 import { StSelectModule } from '../st-select/st-select.module';
+import { StSwitchModule } from '../st-switch/st-switch.module';
 
 @NgModule({
    imports: [
@@ -28,6 +29,7 @@ import { StSelectModule } from '../st-select/st-select.module';
       StInputModule,
       PipesModule,
       StSelectModule,
+      StSwitchModule,
       StFormDirectiveModule
    ],
    declarations: [StFormComponent],